pyproject.toml
[project]
name = "pmpcontrol"
version = "0.0.1"
authors = [
  { name="Kay0n", email="kayon5555@gmail.com" },
]
description = "A Python interface for interacting with the iCON Platform M+ MIDI control surface. "
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "python-rtmidi",
]


[project.urls]
Homepage = "https://github.com/Kay0n/pmpcontrol"
Issues = "https://github.com/Kay0n/pmpcontrol/issues"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

